Update inst/prod for new netboot evolutions
With recent evolutions in the network boot components in Scibian HPC
stack, the § Internal configuration repository > Initial files of the
Installation procedure becomes useless. Then it is removed with this
commit along with its example files.

At the same time, Production procedures are added to explain how to
tune the behaviour of the new components. References to these
procedures is added in the installation procedure in case of
installation issues.